In createRepositoy, Location was not configurable => I added it Also Title was not configurable => I added it In the demo I also added the title for automated repo creation Recent Packages were not updated => run "ncu" and "ncu -u" and "npm i" In tests no repo id and location was passed => I added it Not all files were using tabs but spaces => I changed it Not all files are formatted correctly => please always use format document before pushing! Deleterepo was still deleterepo => deleteRepository (see naming conventions)
execRequest in deleteRepository is asynchronous. you missed an "await" here to get a proper error result => I added await and interpret the result.

The GraphDB Admin Package adds its function to the GraphDB Client Class, so all features of the admin can be used by using the client. You used EnapsoGraphDBAdmin, this was the reason why getBaseURL() was not found. Please check the demo.js of the enapso-graphdb-admin package. Here are a lot of examples.